Refactor feature:write
