+++ app/build.gradle
@@ -22,6 +22,7 @@ android {
         vectorDrawables {
             useSupportLibrary true
         }
+        multiDexEnabled true
     }
 
     buildTypes {
@@ -53,69 +54,29 @@ android {
 
 dependencies {
 
-    implementation 'androidx.core:core-ktx:1.9.0'
-    implementation 'androidx.lifecycle:lifecycle-runtime-ktx:2.6.1'
-    implementation 'androidx.activity:activity-compose:1.7.0'
-    implementation "androidx.compose.ui:ui:$compose_version"
-    implementation "androidx.compose.ui:ui-tooling-preview:$compose_version"
-    implementation 'androidx.compose.material3:material3:1.1.0-beta01'
-    testImplementation 'junit:junit:4.13.2'
-    androidTestImplementation 'androidx.test.ext:junit:1.1.5'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
-    androidTestImplementation "androidx.compose.ui:ui-test-junit4:$compose_version"
-    debugImplementation "androidx.compose.ui:ui-tooling:$compose_version"
-    debugImplementation "androidx.compose.ui:ui-test-manifest:$compose_version"
-
     // Compose Navigation
-    implementation "androidx.navigation:navigation-compose:2.5.3"
+    implementation libs.navigation.compose
 
     // Firebase
-    implementation 'com.google.firebase:firebase-auth-ktx:21.1.0'
-    implementation 'com.google.firebase:firebase-storage-ktx:20.1.0'
+    implementation libs.firebase.storage
 
     // Room components
-    implementation "androidx.room:room-runtime:2.5.1"
-    kapt "androidx.room:room-compiler:2.5.1"
-    implementation "androidx.room:room-ktx:2.5.1"
-
-    // Runtime Compose
-    implementation "androidx.lifecycle:lifecycle-runtime-compose:2.6.1"
+    implementation libs.room.runtime
+    kapt libs.room.compiler
+    implementation libs.room.ktx
 
     // Splash API
-    implementation "androidx.core:core-splashscreen:1.0.0"
+    implementation libs.splash.api
 
     // Mongo DB Realm
-    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-core:1.6.4"
-    implementation "io.realm.kotlin:library-sync:1.7.0"
+    implementation libs.realm.sync
 
     // Dagger Hilt
-    implementation "com.google.dagger:hilt-android:2.44.2"
-    kapt "com.google.dagger:hilt-compiler:2.44.2"
-    implementation "androidx.hilt:hilt-navigation-compose:1.0.0"
-
-    // Coil
-    implementation "io.coil-kt:coil-compose:2.2.2"
-
-    // Pager - Accompanist
-    implementation "com.google.accompanist:accompanist-pager:0.27.0"
-
-    // Date-Time Picker
-    implementation 'com.maxkeppeler.sheets-compose-dialogs:core:1.1.1'
-
-    // CALENDAR
-    implementation "com.maxkeppeler.sheets-compose-dialogs:calendar:1.1.1"
-
-    // CLOCK
-    implementation "com.maxkeppeler.sheets-compose-dialogs:clock:1.1.1"
-
-    // Message Bar Compose
-    implementation 'com.github.stevdza-san:MessageBarCompose:1.0.5'
-
-    // One-Tap Compose
-    implementation 'com.github.stevdza-san:OneTapCompose:1.0.0'
+    implementation libs.hilt.android
+    kapt libs.hilt.compiler
 
     // Desugar JDK
-    coreLibraryDesugaring 'com.android.tools:desugar_jdk_libs:1.2.0'
+    coreLibraryDesugaring libs.desugar.jdk
 
     implementation(project(":core:ui"))
     implementation(project(":core:translations"))
@@ -123,4 +84,5 @@ dependencies {
     implementation(project(":data:mongo"))
     implementation(project(":feature:auth"))
     implementation(project(":feature:home"))
+    implementation(project(":feature:write"))
 }
