+++ app/src/main/java/com/xdiach/diarymoodapp/navigation/NavGraph.kt
@@ -2,27 +2,13 @@
 
 package com.xdiach.diarymoodapp.navigation
 
-import android.widget.Toast
 import androidx.compose.runtime.*
-import androidx.compose.ui.platform.LocalContext
-import androidx.hilt.navigation.compose.hiltViewModel
-import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavHostController
-import androidx.navigation.NavType
 import androidx.navigation.compose.NavHost
-import androidx.navigation.compose.composable
-import androidx.navigation.navArgument
-import com.google.accompanist.pager.ExperimentalPagerApi
-import com.google.accompanist.pager.rememberPagerState
 import com.xdiach.auth.navigation.authenticationRoute
-import com.xdiach.translations.R
-import com.xdiach.util.model.Mood
-import com.xdiach.diarymoodapp.presentation.screens.write.WriteScreen
-import com.xdiach.diarymoodapp.presentation.screens.write.WriteViewModel
 import com.xdiach.home.navigation.homeRoute
-import com.xdiach.ui.UiText
-import com.xdiach.util.Constants.WRITE_SCREEN_ARGUMENT_KEY
 import com.xdiach.util.Screen
+import com.xdiach.write.navigation.writeRoute
 
 @Composable
 fun SetupNavGraph(
@@ -58,88 +44,6 @@ fun SetupNavGraph(
             onBackPressed = {
                 navController.popBackStack()
             }
-        )
-    }
-}
-
-@OptIn(ExperimentalPagerApi::class)
-fun NavGraphBuilder.writeRoute(
-    onBackPressed: () -> Unit
-) {
-    composable(
-        route = Screen.Write.route,
-        arguments = listOf(
-            navArgument(name = WRITE_SCREEN_ARGUMENT_KEY) {
-                type = NavType.StringType
-                nullable = true
-                defaultValue = null
-            }
-        )
-    ) {
-        val viewModel: WriteViewModel = hiltViewModel()
-        val context = LocalContext.current
-        val uiState = viewModel.uiState
-        val pagerState = rememberPagerState()
-        val galleryState = viewModel.galleryState
-        val pageNumber by remember {
-            derivedStateOf { pagerState.currentPage }
-        }
-
-        WriteScreen(
-            uiState = uiState,
-            moodName = { Mood.values()[pageNumber].name },
-            pagerState = pagerState,
-            galleryState = galleryState,
-            onTitleChanged = {
-                viewModel.setTitle(title = it)
-            },
-            onDescriptionChanged = {
-                viewModel.setDescription(description = it)
-            },
-            onBackPressed = onBackPressed,
-            onDeleteConfirmed = {
-                viewModel.deleteDiary(
-                    onSuccess = {
-                        Toast.makeText(
-                            context,
-                            UiText.StringResource(R.string.write_diary_delete_success).asString(
-                                context
-                            ),
-                            Toast.LENGTH_SHORT
-                        ).show()
-                        onBackPressed()
-                    },
-                    onError = { message ->
-                        Toast.makeText(
-                            context,
-                            message,
-                            Toast.LENGTH_SHORT
-                        ).show()
-                    }
-                )
-            },
-            onDateTimeUpdated = { viewModel.updateDateTime(zonedDateTime = it) },
-            onSaveClicked = {
-                viewModel.upsertDiary(
-                    diary = it.apply { mood = Mood.values()[pageNumber].name },
-                    onSuccess = { onBackPressed() },
-                    onError = { message ->
-                        Toast.makeText(
-                            context,
-                            message,
-                            Toast.LENGTH_SHORT
-                        ).show()
-                    }
-                )
-            },
-            onImageSelect = {
-                val type = context.contentResolver.getType(it)?.split("/")?.last() ?: "jpg"
-                viewModel.addImage(
-                    image = it,
-                    imageType = type
-                )
-            },
-            onImageDeleteClicked = { galleryState.removeImage(it) }
         )
     }
 }
