+++ feature/write/build.gradle.kts
@@ -1,6 +1,9 @@
 plugins {
     id("com.android.library")
     id("org.jetbrains.kotlin.android")
+    id("io.realm.kotlin")
+    id("kotlin-kapt")
+    id("dagger.hilt.android.plugin")
 }
 
 android {
@@ -43,4 +46,27 @@ dependencies {
     implementation(libs.activity.compose)
     implementation(libs.material3.compose)
     implementation(libs.navigation.compose)
+
+    implementation(libs.coil)
+
+    implementation(libs.accompanist.pager)
+
+    implementation(libs.coroutines.core)
+    implementation(libs.realm.sync)
+
+    implementation(libs.date.time.picker)
+    implementation(libs.date.dialog)
+    implementation(libs.time.dialog)
+
+    implementation(libs.hilt.android)
+    kapt(libs.hilt.compiler)
+    implementation(libs.hilt.navigation.compose)
+
+    implementation(libs.firebase.auth)
+    implementation(libs.firebase.storage)
+
+    implementation(project(":core:translations"))
+    implementation(project(":core:ui"))
+    implementation(project(":core:util"))
+    implementation(project(":data:mongo"))
 }
