+++ feature/write/src/main/java/com/xdiach/write/navigation/WriteNavigation.kt
@@ -1,0 +1,103 @@
+package com.xdiach.write.navigation
+
+import android.widget.Toast
+import androidx.compose.runtime.derivedStateOf
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.remember
+import androidx.compose.ui.platform.LocalContext
+import androidx.hilt.navigation.compose.hiltViewModel
+import androidx.navigation.NavGraphBuilder
+import androidx.navigation.NavType
+import androidx.navigation.compose.composable
+import androidx.navigation.navArgument
+import com.google.accompanist.pager.ExperimentalPagerApi
+import com.google.accompanist.pager.rememberPagerState
+import com.xdiach.translations.R
+import com.xdiach.ui.UiText
+import com.xdiach.util.Constants
+import com.xdiach.util.Screen
+import com.xdiach.util.model.Mood
+import com.xdiach.write.presentation.screen.WriteScreen
+import com.xdiach.write.presentation.viewmodel.WriteViewModel
+
+@OptIn(ExperimentalPagerApi::class)
+fun NavGraphBuilder.writeRoute(
+    onBackPressed: () -> Unit
+) {
+    composable(
+        route = Screen.Write.route,
+        arguments = listOf(
+            navArgument(name = Constants.WRITE_SCREEN_ARGUMENT_KEY) {
+                type = NavType.StringType
+                nullable = true
+                defaultValue = null
+            }
+        )
+    ) {
+        val viewModel: WriteViewModel = hiltViewModel()
+        val context = LocalContext.current
+        val uiState = viewModel.uiState
+        val pagerState = rememberPagerState()
+        val galleryState = viewModel.galleryState
+        val pageNumber by remember {
+            derivedStateOf { pagerState.currentPage }
+        }
+
+        WriteScreen(
+            uiState = uiState,
+            moodName = { Mood.values()[pageNumber].name },
+            pagerState = pagerState,
+            galleryState = galleryState,
+            onTitleChanged = {
+                viewModel.setTitle(title = it)
+            },
+            onDescriptionChanged = {
+                viewModel.setDescription(description = it)
+            },
+            onBackPressed = onBackPressed,
+            onDeleteConfirmed = {
+                viewModel.deleteDiary(
+                    onSuccess = {
+                        Toast.makeText(
+                            context,
+                            UiText.StringResource(R.string.write_diary_delete_success).asString(
+                                context
+                            ),
+                            Toast.LENGTH_SHORT
+                        ).show()
+                        onBackPressed()
+                    },
+                    onError = { message ->
+                        Toast.makeText(
+                            context,
+                            message,
+                            Toast.LENGTH_SHORT
+                        ).show()
+                    }
+                )
+            },
+            onDateTimeUpdated = { viewModel.updateDateTime(zonedDateTime = it) },
+            onSaveClicked = {
+                viewModel.upsertDiary(
+                    diary = it.apply { mood = Mood.values()[pageNumber].name },
+                    onSuccess = { onBackPressed() },
+                    onError = { message ->
+                        Toast.makeText(
+                            context,
+                            message,
+                            Toast.LENGTH_SHORT
+                        ).show()
+                    }
+                )
+            },
+            onImageSelect = {
+                val type = context.contentResolver.getType(it)?.split("/")?.last() ?: "jpg"
+                viewModel.addImage(
+                    image = it,
+                    imageType = type
+                )
+            },
+            onImageDeleteClicked = { galleryState.removeImage(it) }
+        )
+    }
+}
