+++ feature/write/src/main/java/com/xdiach/write/presentation/components/WriteTopBar.kt
@@ -1,4 +1,4 @@
-package com.xdiach.diarymoodapp.presentation.screens.write
+package com.xdiach.write.presentation.components
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxWidth
